stress-sysfs: make variables i and n plain signed ints.
Clean up cppcheck style warning:
stress-sysfs.c:583:8: style: Checking if unsigned expression 'n' is
less than zero. [unsignedLessThanZero]
 if (n <= 0) {

+++ stress-sysfs.c
@@ -573,7 +573,7 @@ dt_dir_free:
  */
 static int stress_sysfs(const stress_args_t *args)
 {
-	size_t i, n;
+	int i, n;
 	pthread_t pthreads[MAX_SYSFS_THREADS];
 	int rc, ret[MAX_SYSFS_THREADS];
 	stress_ctxt_t ctxt;
